Merge commit '0d4f9f105c3a31475a36e9a1edfa868c66ff455f' into gingerbread-plus-aosp

* commit '0d4f9f105c3a31475a36e9a1edfa868c66ff455f':
  Get wifi limit from secure settings.

+++ tests/src/com/android/providers/downloads/PublicApiFunctionalTest.java
@@ -277,7 +277,7 @@ public class PublicApiFunctionalTest extends AbstractPublicApiTest {
     }
 
     public void testSizeLimitOverMobile() throws Exception {
-        mSystemFacade.mMaxBytesOverMobile = FILE_CONTENT.length() - 1;
+        mSystemFacade.mMaxBytesOverMobile = (long) FILE_CONTENT.length() - 1;
 
         mSystemFacade.mActiveNetworkType = ConnectivityManager.TYPE_MOBILE;
         enqueueResponse(HTTP_OK, FILE_CONTENT);
